+++ src/main/scala/com/trafficland/augmentsbt/releasemanagement/FinalReleaseTasks.scala
@@ -5,42 +5,42 @@ import sbt.{Setting, SettingKey}
 object FinalReleaseTasks {
 
   lazy val releasePublishLibFinalSettingKey: SettingKey[Seq[String]] = SettingKey[Seq[String]] (
-    "release-publish-lib-final-tasks",
+    "releasePublishLibFinalTasks",
     "a list of tasks to execute (in order) for publishing a library's final release"
   )
 
   lazy val releasePublishLibFinalTasks: Setting[Seq[String]] = releasePublishLibFinalSettingKey := Seq(
-    "release-ready",
+    "releaseReady",
     "versionToFinal",
-    "+publish-local",
+    "+publishLocal",
     "+publish",
-    "git-release-commit",
-    "git-checkout-master",
-    "git-merge-develop",
-    "git-tag",
-    "git-checkout-develop",
+    "gitReleaseCommit",
+    "gitCheckoutMaster",
+    "gitMergeDevelop",
+    "gitTag",
+    "gitCheckoutDevelop",
     "versionBumpPatch",
     "versionToSnapshot",
-    "git-version-bump-commit",
-    "git-push-origin"
+    "gitVersionBumpCommit",
+    "gitPushOrigin"
   )
 
   lazy val releaseAppFinalSettingKey: SettingKey[Seq[String]] = SettingKey[Seq[String]] (
-    "release-app-final-tasks",
+    "releaseAppFinalTasks",
     "a list of tasks to execute (in order) for releasing an app's final release"
   )
 
   lazy val releaseAppFinalTasks: Setting[Seq[String]] = releaseAppFinalSettingKey := Seq(
-    "release-ready",
+    "releaseReady",
     "versionToFinal",
-    "git-release-commit",
-    "git-checkout-master",
-    "git-merge-develop",
-    "git-tag",
-    "git-checkout-develop",
+    "gitReleaseCommit",
+    "gitCheckoutMaster",
+    "gitMergeDevelop",
+    "gitTag",
+    "gitCheckoutDevelop",
     "versionBumpPatch",
     "versionToSnapshot",
-    "git-version-bump-commit",
-    "git-push-origin"
+    "gitVersionBumpCommit",
+    "gitPushOrigin"
   )
 }