+++ src/main/scala/com/trafficland/augmentsbt/releasemanagement/SnapshotReleaseTasks.scala
@@ -5,34 +5,34 @@ import sbt.{Setting, SettingKey}
 object SnapshotReleaseTasks {
 
   lazy val releasePublishLibSnapshotSettingKey: SettingKey[Seq[String]] = SettingKey[Seq[String]] (
-    "release-publish-lib-snapshot-tasks",
+    "releasePublishLibSnapshotTasks",
     "a list of tasks to execute (in order) for publishing a library's snapshot release"
   )
 
   lazy val releasePublishLibSnapshotTasks: Setting[Seq[String]] = releasePublishLibSnapshotSettingKey := Seq(
-    "release-ready",
-    "+publish-local",
+    "releaseReady",
+    "+publishLocal",
     "+publish",
     "versionWriteSnapshotRelease",
-    "git-release-commit",
-    "git-tag",
+    "gitReleaseCommit",
+    "gitTag",
     "versionToSnapshot",
-    "git-version-bump-commit",
-    "git-push-origin"
+    "gitVersionBumpCommit",
+    "gitPushOrigin"
   )
 
   lazy val releaseAppSnapshotSettingKey: SettingKey[Seq[String]] = SettingKey[Seq[String]] (
-    "release-app-snapshot-tasks",
+    "releaseAppSnapshotTasks",
     "a list of tasks to execute (in order) for releasing an app's snapshot release"
   )
 
   lazy val releaseAppSnapshotTasks: Setting[Seq[String]] = releaseAppSnapshotSettingKey := Seq(
-    "release-ready",
+    "releaseReady",
     "versionWriteSnapshotRelease",
-    "git-release-commit",
-    "git-tag",
+    "gitReleaseCommit",
+    "gitTag",
     "versionToSnapshot",
-    "git-version-bump-commit",
-    "git-push-origin"
+    "gitVersionBumpCommit",
+    "gitPushOrigin"
   )
 }