+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 1.2.0
 + Upgrade to SBT 1.0 and Scala 2.12
++ Remove key generation.
++ Move tasks to camelCase.
 
 ### 1.1.3
 + Remove unneeded PackageManagementPlugin.